requirements.txt
google-cloud-storage==2.19.0
requests==2.32.3
